Your commit message

+++ llava/data/download_images.py
@@ -1,74 +1,13 @@
-# import os
-# import json
-# import shutil
-# import tarfile
-# import argparse
-# import logging
-# import urllib.request
-# from urllib.error import HTTPError, URLError
-# from tqdm import tqdm
+# ORGINAL CODE THAT WORKED 
 
-# def download_file(url, output_path, max_retries=3):
-#     retries = 0
-#     while retries < max_retries:
-#         try:
-#             urllib.request.urlretrieve(url, output_path)
-#             return True
-#         except (HTTPError, URLError) as e:
-#             logging.warning(f"Retry {retries+1}: Error downloading {url} - {e}")
-#             retries += 1
-#     return False
 
-# def main(args):
-#     input_data = []
-#     with open(args.input_path) as f:
-#         for line in f:
-#             input_data.append(json.loads(line))
-
-#     # Download all PMC articles
-#     logging.info('Downloading PMC articles')
-#     for sample in tqdm(input_data):
-#         pmc_tar_url = sample['pmc_tar_url']
-#         output_path = os.path.join(args.pmc_output_path, os.path.basename(pmc_tar_url))
-#         if not download_file(pmc_tar_url, output_path):
-#             logging.error(f'Failed to download {pmc_tar_url} after retries')
-
-#     # Untar all PMC articles
-#     logging.info('Untarring PMC articles')
-#     for sample in tqdm(input_data):
-#         fname = os.path.join(args.pmc_output_path, os.path.basename(sample['pmc_tar_url']))
-#         if os.path.exists(fname):
-#             try:
-#                 with tarfile.open(fname, "r:gz") as tar:
-#                     tar.extractall(args.pmc_output_path)
-#             except tarfile.TarError as e:
-#                 logging.error(f"Error extracting {fname}: {e}")
-
-#     # Copy to images directory
-#     logging.info('Copying images')
-#     for sample in tqdm(input_data):
-#         src = os.path.join(args.pmc_output_path, sample['image_file_path'])
-#         dst = os.path.join(args.images_output_path, sample['pair_id']+'.jpg')
-#         if os.path.exists(src) and os.path.isfile(src):
-#             shutil.copyfile(src, dst)
-#         else:
-#             logging.warning(f"Image file not found or invalid: {src}")
-
-# if __name__ == '__main__':
-#     logging.basicConfig(level=logging.INFO)
-#     parser = argparse.ArgumentParser()
-#     parser.add_argument('--input_path', type=str, default='data/llava_med_image_urls.jsonl')
-#     parser.add_argument('--pmc_output_path', type=str, default='data/pmc_articles/')
-#     parser.add_argument('--images_output_path', type=str, default='data/images/')
-#     args = parser.parse_args()
-#     main(args)
 
 # import os
 # import json
 # import shutil
 # from tqdm import tqdm
 # import tarfile
-# import argparse
+# import argparse 
 # from urllib.error import HTTPError
 
 # import urllib.request
@@ -95,6 +34,7 @@
 #         for future in tqdm(as_completed(results), total=len(input_data)):
 #             future.result()
 
+
 #     # Untar all PMC articles
 #     print('Untarring PMC articles')
 #     for sample in tqdm(input_data):
@@ -102,7 +42,7 @@
 #         tar = tarfile.open(fname, "r:gz")
 #         tar.extractall(args.pmc_output_path)
 #         tar.close()
-        
+
 #     # Copy to images directory
 #     print('Copying images')
 #     for sample in tqdm(input_data):
@@ -120,52 +60,52 @@
 #     main(args)
 
 
+
+# NOW THAT the download is done focusing on the untarring, untaring one PMC article one by one and then copying to the images path 
+
 import os
 import json
 import shutil
+from tqdm import tqdm
 import tarfile
 import argparse
 from urllib.error import HTTPError
+
 import urllib.request
-from multiprocessing import Pool
-import multiprocessing as mp
+from concurrent.futures import ThreadPoolExecutor, as_completed
 
-parser = argparse.ArgumentParser()
-parser.add_argument('--input_path', type=str, default='data/llava_med_image_urls.jsonl')
-parser.add_argument('--pmc_output_path', type=str, default='data/pmc_articles/')
-parser.add_argument('--images_output_path', type=str, default='data/images/')
-parser.add_argument('--remove_pmc', action='store_true', default=False, help='remove pmc articles after image extraction')
-parser.add_argument('--cpus', type=int, default=-1, help='number of cpus to use in multiprocessing (default: all)')
-args = parser.parse_args()
+def main(args):
+    input_data = []
+    with open(args.input_path) as f:
+        for line in f:
+            input_data.append(json.loads(line))
 
-input_data = []
-with open(args.input_path) as f:
-    for line in f:
-        input_data.append(json.loads(line))
-
-def download_func(idx):
-    sample = input_data[idx]
-    try:
-        urllib.request.urlretrieve(sample['pmc_tar_url'], os.path.join(args.pmc_output_path, os.path.basename(sample['pmc_tar_url'])))
+    print('Processing PMC articles')
+    for sample in tqdm(input_data):
         fname = os.path.join(args.pmc_output_path, os.path.basename(os.path.join(sample['pmc_tar_url'])))
-        
-        tar = tarfile.open(fname, "r:gz")
-        tar.extractall(args.pmc_output_path)
-        tar.close()
-        src = os.path.join(args.pmc_output_path, sample['image_file_path'])
-        dst = os.path.join(args.images_output_path, sample['pair_id']+'.jpg')
-        shutil.copyfile(src, dst)  
-        if args.remove_pmc:
-            os.remove(fname)
-            shutil.rmtree(os.path.join(args.pmc_output_path, str(os.path.basename(sample['pmc_tar_url']))).split('.tar.gz')[0]+'/')
-    except Exception as e:
-        print(e)
-        
-if args.cpus == -1:
-    cpus = mp.cpu_count()
-else:
-    cpus = args.cpus
-    
-pool = Pool(cpus)
+        # Check if the tar file exists before attempting to untar
+        if os.path.exists(fname):
+            try:
+                tar = tarfile.open(fname, "r:gz")
+                tar.extractall(args.pmc_output_path)
+                tar.close()
 
-pool.map(download_func, range(0, len(input_data)))
+                # After untarring, immediately attempt to copy the image
+                src = os.path.join(args.pmc_output_path, sample['image_file_path'])
+                dst = os.path.join(args.images_output_path, sample['pair_id']+'.jpg')
+                if os.path.exists(src):
+                    shutil.copyfile(src, dst)
+                else:
+                    print(f"Source file not found, skipping copy: {src}")
+            except Exception as e:
+                print(f"Error processing {fname}: {e}")
+        else:
+            print(f"File not found, skipping: {fname}")
+
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--input_path', type=str, default='data/llava_med_image_urls.jsonl')
+    parser.add_argument('--pmc_output_path', type=str, default='data/pmc_articles/')
+    parser.add_argument('--images_output_path', type=str, default='data/images/')
+    args = parser.parse_args()
+    main(args)
